+++ NEWS.md
@@ -1,3 +1,11 @@
+# BcStats 0.3.2
+
+* Fixed dependencies
+* Updated column names in calculateSpeed output
+* Updated animal sex modification
+* Fixed bugs in plotting
+
+
 # BcStats 0.3.1
 
 * Fixed dependencies